requirements.txt
streamlit
unstructured
langchain
fastapi
langchain_google_genai
langserve
uvicorn
python-dotenv
sentence-transformers
playwright
selenium
webdriver-manager